requirements.txt
extcolors
Pillow
opencv-python
numpy
matplotlib